Final commit for now
I am starting to get burnt out, and still have UI and a lot of optimization.  I will likely use this project as a template for when I move it to Java or C++, but as of right now, Unity is too slow for a project like this.  Thank you for anyone(if there even is) who was here for the journey!